pyproject.toml
[tool.poetry]
name = "my-business"
version = "0.1.0"
description = ""
authors = ["Hugo Chavero <hugochavero@gmail.com>"]
readme = "README.md"
# packages = [{include = "my_business"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
django-filter = "^23.1"
ipython = "^8.13.1"
black = "^23.7.0"
pylint = "^2.17.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
